refactor: modularize room and profile logic
