+++ components/app/HomePageInner.tsx
@@ -3,7 +3,7 @@
 
 import { useEffect, useRef, useState } from 'react'
 import { useBroadcastEvent, useEventListener } from '@liveblocks/react'
-import { useRouter, usePathname } from 'next/navigation'
+import { useRouter } from 'next/navigation'
 import CharacterSheet, { defaultPerso } from '@/components/sheet/CharacterSheet'
 import DiceRoller from '@/components/dice/DiceRoller'
 import ChatBox from '@/components/chat/ChatBox'
@@ -15,12 +15,14 @@ import SideNotes from '@/components/misc/SideNotes'
 import Login from '@/components/login/Login'
 import GMCharacterSelector from '@/components/misc/GMCharacterSelector'
 import ImportExportMenu from '@/components/character/ImportExportMenu'
+import useDiceHistory from './hooks/useDiceHistory'
+import useProfile from './hooks/useProfile'
+import useOnlineStatus from './hooks/useOnlineStatus'
 
 export default function HomePageInner() {
   const router = useRouter()
-  const pathname = usePathname()
   const [user, setUser] = useState<string | null>(null)
-  const [profile, setProfile] = useState<{ pseudo: string; color: string; isMJ: boolean }>({ pseudo: '', color: '#ffffff', isMJ: false })
+  const profile = useProfile()
   const [perso, setPerso] = useState(defaultPerso)
   const [characters, setCharacters] = useState<any[]>([])
 
@@ -29,7 +31,7 @@ export default function HomePageInner() {
   const [diceResult, setDiceResult] = useState<number | null>(null)
   const [diceDisabled, setDiceDisabled] = useState(false)
   const [pendingRoll, setPendingRoll] = useState<{ result: number; dice: number; nom: string } | null>(null)
-  const [history, setHistory] = useState<{ player: string; dice: number; result: number }[]>([])
+  const [history, setHistory] = useDiceHistory()
   const chatBoxRef = useRef<HTMLDivElement>(null)
 
   const broadcast = useBroadcastEvent()
@@ -42,31 +44,6 @@ export default function HomePageInner() {
     }
   })
 
-  const HISTORY_KEY = 'jdr_dice_history'
-  const lastLength = useRef(0)
-
-  const restoreHistory = () => {
-    try {
-      const saved = localStorage.getItem(HISTORY_KEY)
-      if (saved) {
-        const arr = JSON.parse(saved)
-        if (!Array.isArray(arr)) return
-        if (arr.length !== lastLength.current) {
-          setHistory(arr)
-          lastLength.current = arr.length
-        }
-      }
-    } catch {}
-  }
-
-  useEffect(() => {
-    restoreHistory()
-  }, [pathname])
-
-  useEffect(() => {
-    localStorage.setItem(HISTORY_KEY, JSON.stringify(history))
-    lastLength.current = history.length
-  }, [history])
 
   useEffect(() => {
     if (typeof window === 'undefined') return
@@ -76,66 +53,11 @@ export default function HomePageInner() {
     }
   }, [router])
 
-  useEffect(() => {
-    if (!user) return
-    const id = localStorage.getItem('jdr_profile_id') || crypto.randomUUID()
-    localStorage.setItem('jdr_profile_id', id)
-    const updateOnline = () => {
-      try {
-        const list = JSON.parse(localStorage.getItem('jdr_online') || '{}')
-        list[id] = { pseudo: user, color: profile.color }
-        localStorage.setItem('jdr_online', JSON.stringify(list))
-        window.dispatchEvent(new Event('jdr_online_change'))
-      } catch {}
-    }
-    updateOnline()
-    const handleUnload = () => {
-      try {
-        const list = JSON.parse(localStorage.getItem('jdr_online') || '{}')
-        delete list[id]
-        localStorage.setItem('jdr_online', JSON.stringify(list))
-        window.dispatchEvent(new Event('jdr_online_change'))
-      } catch {}
-    }
-    window.addEventListener('beforeunload', handleUnload)
-    return () => {
-      handleUnload()
-      window.removeEventListener('beforeunload', handleUnload)
-    }
-  }, [user, profile.color])
+  useOnlineStatus(user, profile)
 
   useEffect(() => {
-    try {
-      const saved = localStorage.getItem('jdr_profile')
-      if (saved) {
-        const p = JSON.parse(saved)
-        if (p.pseudo && p.loggedIn) {
-          setUser(p.pseudo)
-          setProfile({ pseudo: p.pseudo, color: p.color || '#ffffff', isMJ: !!p.isMJ })
-        }
-      }
-    } catch {}
-  }, [])
-
-  useEffect(() => {
-    const update = () => {
-      try {
-        const saved = localStorage.getItem('jdr_profile')
-        if (saved) {
-          const p = JSON.parse(saved)
-          if (p.loggedIn) {
-            setProfile({ pseudo: p.pseudo || '', color: p.color || '#ffffff', isMJ: !!p.isMJ })
-          }
-        }
-      } catch {}
-    }
-    window.addEventListener('storage', update)
-    window.addEventListener('jdr_profile_change', update as EventListener)
-    return () => {
-      window.removeEventListener('storage', update)
-      window.removeEventListener('jdr_profile_change', update as EventListener)
-    }
-  }, [])
+    if (profile) setUser(profile.pseudo)
+  }, [profile])
 
   useEffect(() => {
     const savedChars = localStorage.getItem('jdr_characters')
@@ -209,7 +131,7 @@ export default function HomePageInner() {
       <div className="relative z-10 flex w-full h-full">
         <CharacterSheet perso={perso} onUpdate={handleUpdatePerso} chatBoxRef={chatBoxRef} allCharacters={characters} logoOnly>
           <ImportExportMenu perso={perso} onUpdate={handleUpdatePerso} />
-          {profile.isMJ && (
+          {profile?.isMJ && (
             <span className="ml-2">
               <GMCharacterSelector onSelect={handleUpdatePerso} className="bg-gray-800 hover:bg-gray-700 text-white p-2 rounded shadow" />
             </span>
